🚚  Rename types folder into typings
This enable the activation of a file icon for this folder. Much prettier 🤩
