+++ src/Conceptor.ts
@@ -8,7 +8,7 @@
 import * as vscode from 'vscode';
 import { CRCCardsGenerator } from './CRCCardsGenerator';
 import { DesignGraphGenerator } from './DesignGraphGenerator';
-import { CRCCard } from './types/model';
+import { CRCCard } from './typings/model';
 import { ConfigurationManager } from './ConfigurationManager';
 import { ConceptorPanelManager } from './ConceptorPanelManager';
