+++ src/typings/model.d.ts
