+++ src/typings/view.d.ts
